refactor: Split dashjs related functions into a separate file
Moved the dashjs related functions into a new file utils/dash/dashjs.kt to better organize the code.
